write_channel/requirements.txt
bottle==0.12.19
kafka-python==2.0.2
psycopg2-binary==2.8.6
